% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_owl_main.R
\name{fit.owl}
\alias{fit.owl}
\title{Fit an Outcome-Weighted Learning (OWL) Model}
\usage{
fit.owl(
  itr_formula,
  prop_formula,
  mean_formula = NULL,
  method,
  trt,
  outcome,
  dat,
  na.rm = T,
  minimize = F,
  center.outcome = F,
  kernel = "linear",
  cross = 10,
  scale = T,
  seed,
  ...
)
}
\arguments{
\item{itr_formula}{A formula specifying the independent variables for treatment rule optimization.}

\item{prop_formula}{A formula specifying the independent variables for propensity score modeling.}

\item{mean_formula}{A formula specifying the independent variables for mean outcome model.}

\item{method}{Choose either "mr" for multiply robust method or "ipw" for inverse probability weighting method}

\item{trt}{A character string specifying the name of the treatment variable in the dataset.}

\item{dat}{A data frame containing the data for the model.}

\item{kernel}{A character string specifying the kernel type for the SVM (default is "linear").}

\item{cross}{An integer specifying the number of folds for cross-validation in the SVM (default is 10).}

\item{scale}{Logical, whether to scale the data for the SVM (default is TRUE).}
}
\value{
A numeric vector of predicted treatment decisions.
}
\description{
This function fits an OWL model using a weighted SVM (wSVM) to optimize individual treatment rules.
}
\examples{
# Example usage
sim_dat = dat.gen.owl(23, nsample = 250)

model.owl = fit.owl(itr_formula = ~ X.1 + X.2 + X.3, prop_formula = ~ X.1 + X.2, mean_formula = ~ A*(X.1 + X.2) - X.1 - X.2, method = "mr", trt = "A", outcome = "Y", dat = sim_dat,
                   center.outcome = F, seed = 98734)

pred.val = predict(model.owl$model, newdata = sim_dat)

dat <- data.frame(Y = rnorm(100), A = rbinom(100, 1, 0.5), X1 = rnorm(100), X2 = rnorm(100))
itr_formula <- ~ X1 + X2
prop_formula <- ~ X1 + X2
fit.owl(itr_formula, prop_formula, trt = "A", dat = dat)
}
