\name{ITR-package}
\alias{ITR-package}
\alias{ITR}
\docType{package}
\title{
  ITR
}
\description{
  The `ITR` package provides a C++ library for the individualized treatment
  recommendation. 
}
\author{
Bo Zhang, Haoda Fu, Jie Xue

Maintainer: Bo Zhang <zhang_bo3@lilly.com>
}
\references{
  Fu, H., Zhou, J., and Faries, D. E. (2016) Estimating optimal treatment regimes
  via subgroup identification in randomized control trials and observational 
  studies. \emph{Statist. Med.}, 35:3285-3302. doi: 10.1002/sim.6920
}
\keyword{ package }
\seealso{
  Development for ITR can be followed via the GitHub repository at 
  \url{https://github.com/EliLillyCo/ITR}. 
}
